requirements.txt.txt
adagio==0.2.4
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.4
altair==4.2.2
ansi2html==1.8.0
antlr4-python3-runtime==4.11.1
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
autoviz==0.1.603
azure-core==1.26.4
azure-storage-blob==12.16.0
backcall==0.2.0
beautifulsoup4==4.12.2
bleach==6.0.0
blinker==1.6.2
bokeh==2.4.3
boto3==1.26.129
botocore==1.29.129
cachetools==5.3.0
catboost==1.2
category-encoders==2.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmaes==0.9.1
codecov==2.1.13
colorama==0.4.6
colorcet==3.0.1
colorlog==6.7.0
colour==0.1.5
comm==0.1.3
contourpy==1.0.7
coverage==7.2.5
cryptography==40.0.2
cycler==0.11.0
Cython==0.29.34
daal==2023.1.1
daal4py==2023.1.1
dash==2.9.3
dash-auth==2.0.0
dash-bootstrap-components==1.4.1
dash-core-components==2.0.0
dash-cytoscape==0.3.0
dash-html-components==2.0.0
dash-table==5.0.0
dash-testing-stub==0.0.2
dask==2023.4.1
databricks-cli==0.17.6
debugpy==1.6.7
decorator==5.1.1
deepchecks==0.13.1
defusedxml==0.7.1
Deprecated==1.2.13
deprecation==2.1.0
dill==0.3.6
distlib==0.3.6
distributed==2023.4.1
docker==6.1.0
dtreeviz==2.2.1
emoji==2.2.0
entrypoints==0.4
evidently==0.3.1
exceptiongroup==1.1.1
explainerdashboard==0.4.2.2
fairlearn==0.7.0
fastapi==0.95.1
fastjsonschema==2.16.3
ffmpy==0.3.0
filelock==3.12.0
Flask==2.3.2
flask-simplelogin==0.1.2
Flask-WTF==1.1.1
fonttools==4.39.3
frozenlist==1.3.3
fs==2.4.16
fsspec==2023.5.0
fugue==0.8.3
fugue-sql-antlr==0.1.6
future==0.18.3
gevent==22.10.2
gitdb==4.0.10
GitPython==3.1.31
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.59.0
gradio==3.28.3
gradio_client==0.2.0
graphviz==0.20.1
greenlet==2.0.2
grpcio==1.51.3
h11==0.14.0
holoviews==1.15.4
htmlmin==0.1.12
httpcore==0.17.0
httplib2==0.22.0
httpx==0.24.0
huggingface-hub==0.14.1
hvplot==0.8.3
hyperopt==0.2.7
idna==3.4
ImageHash==4.3.1
imageio==2.28.1
imbalanced-learn==0.10.1
importlib-metadata==5.2.0
iniconfig==2.0.0
interpret==0.3.2
interpret-core==0.3.2
ipykernel==6.22.0
ipython==7.34.0
ipython-genutils==0.2.0
ipywidgets==7.7.5
isodate==0.6.1
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpickle==3.0.1
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-dash==0.4.2
jupyter-server==1.24.0
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.1.4
kaleido==0.2.1
kiwisolver==1.4.4
kmodes==0.12.2
lazy_loader==0.2
lightgbm==3.3.5
lime==0.2.0.1
linkify-it-py==2.0.2
llvmlite==0.40.0
locket==1.0.0
lxml==4.9.2
m2cgen==0.10.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.3
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==2.0.5
mlflow==1.30.1
mlxtend==0.22.0
moto==4.1.8
msgpack==1.0.5
multidict==6.0.4
multimethod==1.9.1
multiprocess==0.70.14
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numba==0.57.0
numpy==1.23.0
oauthlib==3.2.2
optuna==3.1.1
orjson==3.8.12
outcome==1.2.0
oyaml==1.0
packaging==21.3
pandas==1.5.3
pandas-profiling==3.6.6
pandocfilters==1.5.0
panel==0.14.4
param==1.13.0
parso==0.8.3
partd==1.4.0
patsy==0.5.3
percy==2.0.2
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.0
plotly==5.14.1
plotly-resampler==0.8.3.2
pluggy==1.0.0
pmdarima==2.0.3
prometheus-client==0.16.0
prometheus-flask-exporter==0.22.4
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
psycopg2==2.9.6
py4j==0.10.9.7
pyamg==5.0.0
pyaml==23.5.8
pyarrow==12.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycaret==3.0.0
pycparser==2.21
pyct==0.5.0
pydantic==1.10.7
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.6.0
pynndescent==0.5.10
PyNomaly==0.3.3
pyod==1.0.9
pyOpenSSL==23.1.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.3.1
python-dateutil==2.8.2
python-multipart==0.0.6
python-utils==3.5.2
pytz==2022.7.1
pyviz-comms==2.2.1
PyWavelets==1.4.1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==23.2.1
qpd==0.4.1
qtconsole==5.4.3
QtPy==2.3.1
querystring-parser==1.2.4
ray==2.4.0
regex==2023.5.5
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.23.1
retrying==1.3.4
rsa==4.9
s3transfer==0.6.1
SALib==1.4.7
schemdraw==0.16
scikit-base==0.4.3
scikit-image==0.20.0
scikit-learn==1.2.2
scikit-learn-intelex==2023.1.1
scikit-optimize==0.9.0
scikit-plot==0.3.7
scipy==1.9.3
seaborn==0.12.2
selenium==4.2.0
semantic-version==2.10.0
Send2Trash==1.8.2
shap==0.41.0
six==1.16.0
skope-rules==1.0.1
sktime==0.18.0
slicer==0.0.7
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==1.4.48
sqlglot==12.1.0
sqlparse==0.4.4
starlette==0.26.1
statsforecast==1.5.0
statsmodels==0.13.5
tabulate==0.9.0
tangled-up-in-unicode==0.2.0
tbats==1.1.3
tbb==2021.9.0
tblib==1.7.0
tenacity==8.2.2
tensorboardX==2.6
terminado==0.17.1
textblob==0.17.1
threadpoolctl==3.1.0
tifffile==2023.4.12
tinycss2==1.2.1
tomli==2.0.1
toolz==0.12.0
tornado==6.3.1
tqdm==4.64.1
trace-updater==0.0.9.1
traitlets==5.9.0
treeinterpreter==0.2.3
triad==0.8.7
trio==0.22.0
trio-websocket==0.10.2
tune-sklearn==0.4.5
typeguard==2.13.3
types-PyYAML==6.0.12.9
typing_extensions==4.5.0
tzdata==2023.3
uc-micro-py==1.0.2
umap-learn==0.5.3
uritemplate==4.1.1
urllib3==1.26.15
urllib3-secure-extra==0.1.0
uvicorn==0.22.0
virtualenv==20.21.0
visions==0.7.5
waitress==2.1.2
wcwidth==0.2.6
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.3
widgetsnbextension==3.6.4
wordcloud==1.9.1.1
wrapt==1.15.0
wsproto==1.2.0
WTForms==3.0.1
xgboost==1.7.5
xlrd==2.0.1
xmltodict==0.13.0
xxhash==3.2.0
yarl==1.9.2
ydata-profiling==4.1.2
yellowbrick==1.5
zict==3.0.0
zipp==3.15.0
zope.event==4.6
zope.interface==6.0